WorldsServer added with port 8550 CommonTools updated with showMessageBox, runExternalWebBrowser Authentication added ('F'acebook, 'S'olipsis with WorldsServer or 'f'ixed), uilogin/uioptions panels updated curl+md5wrapper libs added Facebook authentication lib added EntityUID updated from unsigned int to string NodeId format with 1 authent prefix char + authent id EntityUID format with <NodeId:str>_<ObjId:ushort><SubObjId:ushort> RakNetServer -s arg now specify the loaded scene EntityUID PeerModuleRakNet default port updated from 8550 to 8880
git-svn-id: svn://scm.gforge.inria.fr/svnroot/solipsis/trunk@948 53240bf5-9d7a-4d4b-845a-3b2958f72d47

+++ navigatorModuleOnlyNav/NavigatorXMLRPCClient.cpp
@@ -31,7 +31,7 @@ using namespace Solipsis;
 using namespace CommonTools;
 
 //-------------------------------------------------------------------------------------
-NavigatorXMLRPCClient::NavigatorXMLRPCClient(const std::string& host, int port, int verbosity, const std::string& extras)
+NavigatorXMLRPCClient::NavigatorXMLRPCClient(const std::string& host, unsigned short port, int verbosity, const std::string& extras)
 {
     mP2NClient = IP2NClient::createClient(host, port, verbosity, extras);
     mP2NClient->setLogger(this);
