+++ protocols/P2NClient/ulxmlrpcpp/ULXMLRPCPPP2NClient.h
@@ -50,7 +50,7 @@ protected:
     /// Host
     std::string mHost;
     /// Port
-    int mPort;
+    unsigned short mPort;
     /// Verbosity
     int mVerbosity;
     /// Additional informations
@@ -77,13 +77,13 @@ private:
     IP2NClientLogger* mLogger;
 
 public:
-    ULXMLRPCPPP2NClient(const std::string& host, int port, int verbosity, const std::string& extras = "");
+    ULXMLRPCPPP2NClient(const std::string& host, unsigned short port, int verbosity, const std::string& extras = "");
     virtual ~ULXMLRPCPPP2NClient();
 
     /// @copydoc IP2NClient::getHost
     virtual const std::string& getHost() { return mHost; }
     /// @copydoc IP2NClient::getPort
-    virtual int getPort() { return mPort; }
+    virtual unsigned short getPort() { return mPort; }
     /// @copydoc IP2NClient::getVerbosity
     virtual int getVerbosity() { return mVerbosity; }
     /// @copydoc IP2NClient::getExtras
