+++ protocols/P2NNullClientServer/P2NNullClient.h
@@ -37,7 +37,7 @@ protected:
     /// Host
     std::string mHost;
     /// Port
-    int mPort;
+    unsigned short mPort;
     /// Verbosity
     int mVerbosity;
     /// Additional informations
@@ -52,13 +52,13 @@ private:
     IP2NClientLogger* mLogger;
 
 public:
-    P2NNullClient(const std::string& host, int port, int verbosity, const std::string& extras);
+    P2NNullClient(const std::string& host, unsigned short port, int verbosity, const std::string& extras);
     virtual ~P2NNullClient();
 
     /// @copydoc IP2NClient::getHost
     virtual const std::string& getHost() { return mHost; }
     /// @copydoc IP2NClient::getPort
-    virtual int getPort() { return mPort; }
+    virtual unsigned short getPort() { return mPort; }
     /// @copydoc IP2NClient::getVerbosity
     virtual int getVerbosity() { return mVerbosity; }
     /// @copydoc IP2NClient::getExtras
