+++ protocols/P2NNullClientServer/P2NNullServer.h
@@ -39,7 +39,7 @@ private:
     /** Host */
     std::string mHost;
     /** Port Number */
-    int mPort;
+    unsigned short mPort;
     /** Additional informations */
     std::string mExtras;
 
@@ -51,7 +51,7 @@ private:
 
 public:
     /** Constructor */
-    P2NNullServer(IP2NServerRequestsHandler* requestsHandler, const std::string& host = "localhost", int port = 8550, int verbosity = 0, const std::string& extras = "");
+    P2NNullServer(IP2NServerRequestsHandler* requestsHandler, const std::string& host = "localhost", unsigned short port = 8550, int verbosity = 0, const std::string& extras = "");
 
     /** Destructor */
     virtual ~P2NNullServer();
